Moved post endpoints to controller
