+++ internal/ingress/controller/nginx.go
@@ -18,12 +18,10 @@ package controller
 
 import (
 	"bytes"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io/ioutil"
 	"net"
-	"net/http"
 	"os"
 	"os/exec"
 	"strconv"
@@ -337,12 +335,8 @@ func (n *NGINXController) Start() {
 				glog.V(3).Infof("Event %v received - object %v", evt.Type, evt.Obj)
 				if evt.Type == store.ConfigurationEvent {
 					n.SetForceReload(true)
-				} else if evt.Type == store.UpdateEvent {
-					if ep, ok := evt.Obj.(*apiv1.Endpoints); ok {
-						n.SetForceNoReload(true)
-						n.postEndpoints(ep)
-					}
 				}
+
 				n.syncQueue.Enqueue(evt.Obj)
 			} else {
 				glog.Warningf("unexpected event type received %T", event)
@@ -350,45 +344,6 @@ func (n *NGINXController) Start() {
 		case <-n.stopCh:
 			break
 		}
-	}
-}
-
-type Upstream struct {
-	Name string `json:"name"`
-	Down bool   `json:"down"`
-}
-
-// Post endpoints to the HTTDB endpoint for NGINX to read
-func (n *NGINXController) postEndpoints(ep *apiv1.Endpoints) {
-	upstreamPool := []Upstream{}
-
-	for _, ss := range ep.Subsets {
-		for _, epPort := range ss.Ports {
-			// Append addresses to list of upstreams
-			for _, epAddress := range ss.Addresses {
-				upstreamPool = append(upstreamPool, Upstream{Name: fmt.Sprintf("%s:%d", epAddress.IP, epPort.Port), Down: false})
-			}
-		}
-	}
-
-	jsonBody, err := json.Marshal(struct {
-		Key   string     `json:"key"`
-		Value []Upstream `json:"value"`
-		TTL   int        `json:"ttl"`
-	}{
-		"pool-name", //TODO: This needs to be the dynamic Upstream.Name
-		upstreamPool,
-		0,
-	})
-	if err != nil {
-		glog.Warningf("error creating request body %v", err)
-	}
-
-	client := &http.Client{}
-	req, err := http.NewRequest("POST", "http://localhost:18080/httpdb/dicts/dynamic_upstreams/keys", bytes.NewBuffer([]byte(jsonBody)))
-	_, err = client.Do(req)
-	if err != nil {
-		glog.Warningf("error posting endpoint %v", err)
 	}
 }
 
@@ -710,6 +665,11 @@ func (n *NGINXController) OnUpdate(ingressCfg ingress.Configuration) error {
 	err = ioutil.WriteFile(cfgPath, content, 0644)
 	if err != nil {
 		return err
+	}
+
+	if n.isForceNoReload() {
+		glog.V(3).Infof("force no reload, skipping backend reload")
+		return nil
 	}
 
 	o, err := exec.Command(n.binary, "-s", "reload", "-c", cfgPath).CombinedOutput()
